+++ forge-ai/src/main/java/forge/ai/SpellApiToAi.java
@@ -133,7 +133,6 @@ public enum SpellApiToAi {
             .put(ApiType.PlayLandVariant, CannotPlayAi.class)
             .put(ApiType.Poison, PoisonAi.class)
             .put(ApiType.PreventDamage, DamagePreventAi.class)
-            .put(ApiType.PreventDamageAll, DamagePreventAllAi.class)
             .put(ApiType.Proliferate, CountersProliferateAi.class)
             .put(ApiType.Protection, ProtectAi.class)
             .put(ApiType.ProtectionAll, ProtectAllAi.class)
@@ -144,7 +143,6 @@ public enum SpellApiToAi {
             .put(ApiType.Radiation, AlwaysPlayAi.class)
             .put(ApiType.RearrangeTopOfLibrary, RearrangeTopOfLibraryAi.class)
             .put(ApiType.Regenerate, RegenerateAi.class)
-            .put(ApiType.RegenerateAll, RegenerateAllAi.class)
             .put(ApiType.Regeneration, AlwaysPlayAi.class)
             .put(ApiType.RemoveCounter, CountersRemoveAi.class)
             .put(ApiType.RemoveCounterAll, CannotPlayAi.class)
