+++ forge-game/src/main/java/forge/game/ability/ApiType.java
@@ -133,7 +133,6 @@ public enum ApiType {
     PlayLandVariant (PlayLandVariantEffect.class),
     Poison (PoisonEffect.class),
     PreventDamage (DamagePreventEffect.class),
-    PreventDamageAll (DamagePreventAllEffect.class),
     Proliferate (CountersProliferateEffect.class),
     Protection (ProtectEffect.class),
     ProtectionAll (ProtectAllEffect.class),
@@ -144,7 +143,6 @@ public enum ApiType {
     Radiation (RadiationEffect.class),
     RearrangeTopOfLibrary (RearrangeTopOfLibraryEffect.class),
     Regenerate (RegenerateEffect.class),
-    RegenerateAll (RegenerateAllEffect.class),
     Regeneration (RegenerationEffect.class),
     RemoveCounter (CountersRemoveEffect.class),
     RemoveCounterAll (CountersRemoveAllEffect.class),
