+++ forge-game/src/main/java/forge/game/card/CardFactoryUtil.java
@@ -1592,7 +1592,7 @@ public class CardFactoryUtil {
             // replace , for ; in the ChangeZone
             k[1] = k[1].replace(",", ";");
 
-            final String effect = "DB$ ChangeZone | ValidTgts$ Player | TgtPrompt$ Select target player" +
+            final String effect = "DB$ ChangeZone | ValidTgts$ Player" +
                     " | Origin$ Library | Destination$ Hand | ChangeType$ Card.named" + k[1] +
                     " | Hidden$ True | Chooser$ Targeted | Optional$ True";
 
@@ -2898,7 +2898,7 @@ public class CardFactoryUtil {
             abilityStr.append("| ValidTgts$ ").append(valid);
             abilityStr.append(" | TgtPrompt$ Select target ").append(vstr).append(" you control ");
             // the if the Equipment can really attach should be part of the Attach Effect
-            abilityStr.append("| SorcerySpeed$ True | Equip$ True | AILogic$ Pump");
+            abilityStr.append("| SorcerySpeed$ True | AILogic$ Pump");
             // add AttachAi for some special cards
             if (card.hasSVar("AttachAi")) {
                 abilityStr.append("| ").append(card.getSVar("AttachAi"));
