+++ forge-game/src/main/java/forge/game/trigger/WrappedAbility.java
@@ -70,7 +70,6 @@ public class WrappedAbility extends Ability {
             ApiType.DealDamage, // checked
 
             ApiType.Regenerate, // Updated
-            ApiType.RegenerateAll, // No Triggered
             ApiType.Regeneration, // Replacement Effect only
 
             ApiType.DelayedTrigger
