+++ forge-gui/res/cardsfolder/g/golgari_charm.txt
@@ -4,5 +4,5 @@ Types:Instant
 A:SP$ Charm | Choices$ PumpAll,DestroyEnch,RegenAll | CharmNum$ 1
 SVar:PumpAll:DB$ PumpAll | ValidCards$ Creature | NumAtt$ -1 | NumDef$ -1 | IsCurse$ True | SpellDescription$ All creatures get -1/-1 until end of turn.
 SVar:DestroyEnch:DB$ Destroy | ValidTgts$ Enchantment | TgtPrompt$ Select target enchantment | SpellDescription$ Destroy target enchantment.
-SVar:RegenAll:DB$ RegenerateAll | ValidCards$ Creature.YouCtrl | SpellDescription$ Regenerate each creature you control.
+SVar:RegenAll:DB$ Regenerate | Defined$ Valid Creature.YouCtrl | SpellDescription$ Regenerate each creature you control.
 Oracle:Choose one —\n• All creatures get -1/-1 until end of turn.\n• Destroy target enchantment.\n• Regenerate each creature you control.
