+++ forge-gui/res/cardsfolder/w/wail_of_the_nim.txt
@@ -3,6 +3,6 @@ ManaCost:2 B
 Types:Instant
 K:Entwine:B
 A:SP$ Charm | Choices$ DBRegenerate,DBDamageAll | CharmNum$ 1
-SVar:DBRegenerate:DB$ RegenerateAll | ValidCards$ Creature.YouCtrl | SpellDescription$ Regenerate each creature you control.
+SVar:DBRegenerate:DB$ Regenerate | Defined$ Valid Creature.YouCtrl | SpellDescription$ Regenerate each creature you control.
 SVar:DBDamageAll:DB$ DamageAll | ValidCards$ Creature | ValidPlayers$ Player | NumDmg$ 1 | SpellDescription$ CARDNAME deals 1 damage to each creature and each player.
 Oracle:Choose one —\n• Regenerate each creature you control.\n• Wail of the Nim deals 1 damage to each creature and each player.\nEntwine {B} (Choose both if you pay the entwine cost.)
