+++ forge-gui/res/cardsfolder/w/wrap_in_vigor.txt
@@ -1,5 +1,5 @@
 Name:Wrap in Vigor
 ManaCost:1 G
 Types:Instant
-A:SP$ RegenerateAll | ValidCards$ Creature.YouCtrl | SpellDescription$ Regenerate each creature you control.
+A:SP$ RegenerateAll | Defined$ Valid Creature.YouCtrl | SpellDescription$ Regenerate each creature you control.
 Oracle:Regenerate each creature you control.
